rework treasure maps
