+++ locale/en/pirates.cfg
@@ -16,7 +16,7 @@ softmod_info_tips_1=Features of the game that are hard to work out alone
 softmod_info_tips_2=• The captain (or officers) can steer the boat from the crow's nest by placing 50 rail signals in one of the blue boxes.\n• The quantity on an island is independent of the order in which you break rocks.\n• Passive pollution ramps up over time on each island.\n• Items dropped on the deck are transferred to the cabin when the boat moves for performance reasons.\n• Mining productivity increases naturally with leagues traveled.\nOn multiplayer, crews that are inactive for 96 hours are auto-disbanded.\n• Useful commands: /classinfo {classname} command, /plank {player}, /officer {player}, /undock, /ccolor, /clear_north_tanks, /clear_south_tanks, /tax, /reset_password
 
 softmod_info_1_1=v1.7.0
-softmod_info_1_2=• Support for Factorio 2.0.\n• With the introduction of remote view, spectator mode has been retired.\n• 9 simultaneous crews now supported.\n• Performance improvements.
+softmod_info_1_2=• Support for Factorio 2.0.\n• With the introduction of remote view, spectator mode has been retired.\n• 9 simultaneous crews are now supported.\n• The treasure map mechanic has been reworked to use cameras.\n• Performance improvements.
 
 softmod_info_2_1=v1.6.0–1.6.7 highlights
 softmod_info_2_2=• New runs begin on the top lane so that beginners are more likely to encounter docks.\n• Players now die rather than being temporarily marooned if they miss the boat.\n• The pause at sea now occurs after the loading time for the next destination.\n• Optional setting for new crews to disable blueprints.\n• Machines now deactivate after several minutes spent waiting at sea.\n• Private crews and captain-protected crews no longer revert to public crews when inactive. Runs now autodisband after 96 hours of inactivity.\n• Shotguns are now buffed as intended.
@@ -513,6 +513,10 @@ gui_progress_leagues=__1__ leagues
 gui_progress_current_location=Current location: __1__
 gui_progress_time_of_day=Time of day: __1__
 
+gui_treasure=Treasure
+gui_treasure_discovered_maps=Discovered Maps
+gui_treasure_tooltip=The treasure maps the crew has currently found. Dig at the right location to find treasure.
+
 gui_runs_play=Crews
 gui_runs_ongoing_runs=Ongoing Crews
 gui_runs_ongoing_runs_helpful_tip=To join a crew, first select it in the table below.
@@ -567,9 +571,9 @@ gui_etaframe_board_warning_tooltip=Probably time to board...
 gui_etaframe_autoundock=Auto-undock:
 gui_etaframe_loading=Loading
 gui_etaframe_defeat_krakens=Defeat the krakens!
-gui_etaframe_atsea_waiting=Captain — Click here to 
+gui_etaframe_atsea_waiting=Captain — Click here to proceed
 gui_etaframe_atsea_waiting_are_you_sure=Proceed — Are you sure?
-gui_etaframe_atsea_victorious=Captain — Click here to Proceed
+gui_etaframe_atsea_victorious=Captain — Click here to proceed
 gui_etaframe_undock=Undock:
 gui_etaframe_anytime=Anytime
 gui_etaframe_next_escape_cost=Next escape cost:
