+++ maps/pirates/api_on_tick.lua
@@ -530,7 +530,7 @@ function Public.periodic_free_resources(tick_interval)
 	end
 end
 
-function Public.pick_up_tick(tick_interval)
+function Public.pick_up_ghosts_tick(tick_interval)
 	if Common.activecrewcount() == 0 then
 		return
 	end
@@ -549,82 +549,7 @@ function Public.pick_up_tick(tick_interval)
 		return
 	end
 
-	local maps = dynamic_data.treasure_maps or {}
-	local buried_treasure = dynamic_data.buried_treasure or {}
 	local ghosts = dynamic_data.ghosts or {}
-
-	for i = 1, #maps do
-		local map = maps[i]
-
-		if map.state == 'on_ground' then
-			local p = map.position
-
-			local nearby_characters = surface.find_entities_filtered({ position = p, radius = 3, name = 'character' })
-			local nearby_characters_count = #nearby_characters
-			if nearby_characters_count > 0 then
-				local player
-				local j = 1
-				while j <= nearby_characters_count do
-					if
-						nearby_characters[j]
-						and nearby_characters[j].valid
-						and nearby_characters[j].player
-						and Common.validate_player(nearby_characters[j].player)
-					then
-						player = nearby_characters[j].player
-						break
-					end
-					j = j + 1
-				end
-				if player then
-					local buried_treasure_candidates = {}
-					for _, t in pairs(buried_treasure) do
-						if not t.associated_to_map then
-							buried_treasure_candidates[#buried_treasure_candidates + 1] = t
-						end
-					end
-					if #buried_treasure_candidates == 0 then
-						break
-					end
-					local chosen = buried_treasure_candidates[Math.random(#buried_treasure_candidates)]
-
-					chosen.associated_to_map = true
-					local p2 = chosen.position
-					map.buried_treasure_position = p2
-
-					map.state = 'picked_up'
-					map.mapobject_rendering.destroy()
-
-					Common.notify_force_light(player.force, { 'pirates.find_map', player.name })
-
-					map.x_renderings = {
-						rendering.draw_line({
-							width = 8,
-							surface = surface,
-							from = { p2.x + 3, p2.y + 3 },
-							to = { p2.x - 3, p2.y - 3 },
-							color = { 1, 0, 0 },
-							gap_length = 0.2,
-							dash_length = 1,
-							draw_on_ground = true,
-							-- players = {player},
-						}),
-						rendering.draw_line({
-							width = 8,
-							surface = surface,
-							from = { p2.x - 3, p2.y + 3 },
-							to = { p2.x + 3, p2.y - 3 },
-							color = { 1, 0, 0 },
-							gap_length = 0.2,
-							dash_length = 1,
-							draw_on_ground = true,
-							-- players = {player},
-						}),
-					}
-				end
-			end
-		end
-	end
 
 	if not (dynamic_data.quest_type and not dynamic_data.quest_complete) then
 		return
@@ -668,7 +593,7 @@ function Public.pick_up_tick(tick_interval)
 	end
 end
 
-function Public.interpret_shorthanded_force_name(shorthanded_name)
+local function interpret_shorthanded_force_name(shorthanded_name)
 	local memory = Memory.get_crew_memory()
 
 	local ret
@@ -748,7 +673,7 @@ function Public.place_cached_structures(tick_interval)
 
 				local force_name
 				if c.force then
-					force_name = Public.interpret_shorthanded_force_name(c.force)
+					force_name = interpret_shorthanded_force_name(c.force)
 				end
 
 				if c.type == 'tiles' then
@@ -949,131 +874,6 @@ function Public.update_boat_stored_resources(tick_interval)
 	end
 
 	Common.update_boat_stored_resources()
-end
-
-function Public.buried_treasure_check(tick_interval)
-	if Common.activecrewcount() == 0 then
-		return
-	end
-
-	-- local memory = Memory.get_crew_memory()
-	local destination = Common.current_destination()
-
-	local remaining = destination.dynamic_data.treasure_remaining
-
-	if
-		not (
-			remaining
-			and remaining > 0
-			and destination.surface_name
-			and destination.dynamic_data.buried_treasure
-			and #destination.dynamic_data.buried_treasure > 0
-		)
-	then
-		return
-	end
-
-	local surface = game.surfaces[destination.surface_name]
-	local treasure_table = destination.dynamic_data.buried_treasure
-
-	for i = 1, #treasure_table do
-		local treasure = treasure_table[i]
-		if not treasure then
-			break
-		end
-
-		local t = treasure.treasure
-
-		if t then
-			local p = treasure.position
-			local free = surface.can_place_entity({ name = 'wooden-chest', position = p })
-
-			if free then
-				local inserters = {
-					surface.find_entities_filtered({
-						type = 'inserter',
-						position = { x = p.x - 1, y = p.y },
-						radius = 0.1,
-						direction = defines.direction.east,
-					}),
-					surface.find_entities_filtered({
-						type = 'inserter',
-						position = { x = p.x + 1, y = p.y },
-						radius = 0.1,
-						direction = defines.direction.west,
-					}),
-					surface.find_entities_filtered({
-						type = 'inserter',
-						position = { x = p.x, y = p.y - 1 },
-						radius = 0.1,
-						direction = defines.direction.south,
-					}),
-					surface.find_entities_filtered({
-						type = 'inserter',
-						position = { x = p.x, y = p.y + 1 },
-						radius = 0.1,
-						direction = defines.direction.north,
-					}),
-				}
-
-				for j = 1, 4 do
-					if inserters[j] and inserters[j][1] then
-						local ins = inserters[j][1]
-
-						if
-							destination.dynamic_data.treasure_remaining > 0
-							and ins.held_stack.count == 0
-							and ins.status == defines.entity_status.waiting_for_source_items
-						then
-							surface.create_entity({
-								name = 'item-on-ground',
-								position = p,
-								stack = { name = t.name, count = 1 },
-							})
-							t.count = t.count - 1
-							destination.dynamic_data.treasure_remaining = destination.dynamic_data.treasure_remaining
-								- 1
-
-							if destination.dynamic_data.treasure_remaining == 0 then
-								-- destroy all
-								local buried_treasure = destination.dynamic_data.buried_treasure
-								for _, t2 in pairs(buried_treasure) do
-									t2 = nil
-								end
-								local maps = destination.dynamic_data.treasure_maps
-								for _, m in pairs(maps) do
-									if m.state == 'on_ground' then
-										m.mapobject_rendering.destroy()
-									elseif m.state == 'picked_up' and m.x_renderings and #m.x_renderings > 0 then
-										m.x_renderings[1].destroy()
-										m.x_renderings[2].destroy()
-									end
-									m = nil
-								end
-							elseif t.count <= 0 then
-								treasure.treasure = nil
-
-								local maps = destination.dynamic_data.treasure_maps
-								for _, m in pairs(maps) do
-									if
-										m.state == 'picked_up'
-										and m.buried_treasure_position
-										and m.buried_treasure_position == p
-										and m.x_renderings
-										and #m.x_renderings > 0
-									then
-										m.state = 'inactive'
-										m.x_renderings[1].destroy()
-										m.x_renderings[2].destroy()
-									end
-								end
-							end
-						end
-					end
-				end
-			end
-		end
-	end
 end
 
 function Public.boat_movement_tick(tick_interval)
@@ -1989,30 +1789,6 @@ function Public.equalise_fluid_storages()
 			for j = 2, #storages do
 				equalise_fluid_storage_pair(storages[j], storages[j - 1])
 				equalise_fluid_storage_pair(storages[j - 1], storages[j])
-			end
-		end
-	end
-end
-
-function Public.revealed_buried_treasure_distance_check()
-	local destination = Common.current_destination()
-	if Common.activecrewcount() == 0 then
-		return
-	end
-
-	if destination.dynamic_data.some_player_was_close_to_buried_treasure then
-		return
-	end
-
-	local maps = destination.dynamic_data.treasure_maps or {}
-	for _, map in pairs(maps) do
-		if map.state == 'picked_up' then
-			for _, player in pairs(Common.crew_get_crew_members()) do
-				if player.character and player.character.valid then
-					if Math.distance(player.character.position, map.buried_treasure_position) <= 20 then
-						destination.dynamic_data.some_player_was_close_to_buried_treasure = true
-					end
-				end
 			end
 		end
 	end
