+++ maps/pirates/gui/progress.lua
@@ -96,9 +96,6 @@ end
 -- end
 
 function Public.full_update(player)
-	if Public.regular_update then
-		Public.regular_update(player)
-	end
 	if not player.gui.screen[window_name .. '_piratewindow'] then
 		return
 	end
